backend/requirements.txt
# === Web 框架 ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1

# === MQTT ===
paho-mqtt==2.1.0

# === 数据模型与配置 ===
pydantic==2.10.4
pydantic-settings==2.7.1

# === 数据库 ===
sqlalchemy==2.0.36
aiosqlite==0.20.0

# === 工具库 ===
python-dotenv==1.0.1
structlog==24.4.0
crcmod==1.7

# === 测试 ===
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
